refactor(channels): split feishu-channel.ts into modular components (Issue #694)
Split the 1059-line feishu-channel.ts into smaller, focused modules:

- mention-detector.ts (127 lines): Bot mention detection logic
- passive-mode.ts (61 lines): Passive mode state management
- welcome-handler.ts (133 lines): Welcome message event handling
- message-handler.ts (496 lines): Message processing and routing
- card-handler.ts (160 lines): Card action event handling
- index.ts (12 lines): Module exports

Main file reduced from 1059 to 376 lines (~65% reduction).

Benefits:
- Better separation of concerns
- Easier testing of individual components
- Clearer responsibilities
- Improved maintainability

All 36 feishu-channel tests pass.

Fixes #694
