+++ src/channels/feishu/passive-mode.ts
@@ -1,0 +1,61 @@
+/**
+ * Passive Mode Manager for Feishu Channel.
+ *
+ * Manages passive mode state for group chats.
+ *
+ * Issue #511: Group chat passive mode control
+ */
+
+import { createLogger } from '../../utils/logger.js';
+
+const logger = createLogger('PassiveModeManager');
+
+/**
+ * PassiveModeManager - Manages passive mode state for group chats.
+ *
+ * In passive mode, the bot only responds when mentioned (@bot).
+ * This can be disabled per chat to make the bot respond to all messages.
+ */
+export class PassiveModeManager {
+  /**
+   * Passive mode state storage.
+   * Key: chatId, Value: true if passive mode is disabled (bot responds to all messages)
+   */
+  private passiveModeDisabled: Map<string, boolean> = new Map();
+
+  /**
+   * Check if passive mode is disabled for a specific chat.
+   * When passive mode is disabled, the bot responds to all messages in group chats.
+   *
+   * @param chatId - Chat ID to check
+   * @returns true if passive mode is disabled (bot responds to all messages)
+   */
+  isPassiveModeDisabled(chatId: string): boolean {
+    return this.passiveModeDisabled.get(chatId) === true;
+  }
+
+  /**
+   * Set passive mode state for a specific chat.
+   *
+   * @param chatId - Chat ID to configure
+   * @param disabled - true to disable passive mode (respond to all messages)
+   */
+  setPassiveModeDisabled(chatId: string, disabled: boolean): void {
+    if (disabled) {
+      this.passiveModeDisabled.set(chatId, true);
+      logger.info({ chatId }, 'Passive mode disabled for chat');
+    } else {
+      this.passiveModeDisabled.delete(chatId);
+      logger.info({ chatId }, 'Passive mode enabled for chat');
+    }
+  }
+
+  /**
+   * Get all chats with passive mode disabled.
+   *
+   * @returns Array of chat IDs with passive mode disabled
+   */
+  getPassiveModeDisabledChats(): string[] {
+    return Array.from(this.passiveModeDisabled.keys());
+  }
+}
